fix: extract shared test mocks and record code review for issue #189
Extract clipboard and ResizeObserver mock setup into shared
test/mockHelpers.ts to eliminate duplication across 3 test files.
Addresses medium code review finding.
