+++ code-review-results/2026-04-05-issue-189.md
@@ -1,0 +1,57 @@
+# Code Review: Issue #189 — 2026-04-05
+
+## Findings
+
+| Severity | File | Issue | Pillar | Resolved |
+| -------- | ---- | ----- | ------ | -------- |
+| Medium | `frontend/src/components/PatronInfoCard/CopyableField.test.tsx`, `PatronInfoCard.test.tsx`, `useCopyToClipboard.test.tsx` | Clipboard mock setup (`Object.defineProperty(navigator, 'clipboard', ...)`) and `ResizeObserver` mock class are copy-pasted across all 3 test files. CLAUDE.md requires extracting repeated setup/teardown into shared helpers when a block appears 3+ times. | Verification | No |
+| Low | `frontend/src/components/PatronInfoCard/CopyableField.tsx` | Component uses TypeScript interface but does not define PropTypes. CLAUDE.md requires "All React components must define PropTypes." However, the vast majority of .tsx components in this codebase (40+) follow the same pattern of using TS interfaces without PropTypes, making this a systemic gap rather than a PR-specific issue. | Documentation | No |
+| Low | `.claude/settings.json` (hooks) | PostToolUse ESLint and test hooks only match `*.js` and `*.jsx` files. All new files in this PR are `.tsx`/`.ts` and would not be linted or tested by these hooks. This is a pre-existing harness gap. | Verification | No |
+
+## Harness Improvement Recommendations
+
+| Finding | Harness Recommendation | Harness Change Made |
+| ------- | ---------------------- | ------------------- |
+| Repeated test mocks across 3 files | Add a CLAUDE.md constraint or linter check that flags identical beforeEach blocks across test files, or add a shared test helper (e.g., `frontend/src/test-utils/clipboardMock.ts`) | No |
+| PropTypes rule vs TypeScript reality | Update the CLAUDE.md PropTypes convention to clarify that TypeScript interfaces satisfy the prop typing requirement for `.tsx` files, or enforce PropTypes alongside TS interfaces if both are required | No |
+| ESLint hooks miss .ts/.tsx files | Update `.claude/settings.json` hook matchers to include `*.ts` and `*.tsx` patterns: `*/frontend/src/*.ts` and `*/frontend/src/*.tsx` | No |
+
+## Suggested E2E Tests
+
+Based on the changes in this review, the following end-to-end tests are recommended:
+
+| # | Scenario | Steps | Expected Result | Priority |
+|---|----------|-------|-----------------|----------|
+| 1 | Copy email from patron info card | Navigate to patron detail page, click copy button next to email field | Email value copied to clipboard, check icon appears briefly | High |
+| 2 | Copy card number from patron info card | Navigate to patron detail page, click copy button next to card number field | Card number copied to clipboard, check icon appears briefly | High |
+| 3 | Long email truncation with tooltip | Navigate to patron with a very long email address | Email is truncated with ellipsis, hovering shows full email in tooltip | Medium |
+| 4 | Null email displays N/A without copy button | Navigate to patron with no email set | "N/A" displayed, no copy button visible | Medium |
+
+### Notes
+- The project has Playwright smoke tests (`npm run test:smoke`). These scenarios align with existing e2e patterns.
+- Clipboard API access may require browser permissions in Playwright; use `context.grantPermissions(['clipboard-read', 'clipboard-write'])`.
+
+## Harness Self-Audit
+
+Audit source: CLI + inline checks
+
+| Category | Score | Pass | Fail |
+| -------- | ----: | ---: | ---: |
+| Instructions | 81% | 8 | 2 |
+| Verification | 82% | 5 | 1 |
+| Constraints | 100% | 4 | 0 |
+| Context | 75% | 3 | 1 |
+| Tools | 100% | 2 | 0 |
+| Session | 60% | 2 | 2 |
+| Architecture | 50% | 2 | 1 |
+
+**Overall:** 79% (Decent)
+
+### Drift Detected
+
+- 3 of 15 CLAUDE.md path references could not be resolved by CLI audit (likely relative path resolution differences)
+- 31 harness artifacts not reachable from CLAUDE.md (agents, commands, knowledge files not linked from top-level doc)
+- ESLint/test hooks in `.claude/settings.json` do not match `.ts`/`.tsx` files, meaning TypeScript source and test files bypass automated checks
+- PropTypes convention in CLAUDE.md contradicts codebase-wide TypeScript interface pattern (40+ components use TS interfaces without PropTypes)
+- No CI/CD configuration detected
+- No compact guidance found in CLAUDE.md (though `context-management-kit.md` exists)
